// src/emission/enum/purchase-good-and-service.enum.ts
export enum PurchaseGoodAndService{
    supplier_specific_method = 'Supplier Specific Method',
    hybrid_method = 'Hybrid Method',
    average_data_method = 'Average Data Method',
    spend_based_method = 'Spend Based Method'
}

export enum TypeNames{
    purchase = 'Purchased Goods and Services',
    material = 'Material Inputs',
    transport = 'Transport of Material Inputs',
    waste = 'Waste Outputs',
    other = 'Other waste emissions'
}